pkgbuild: cp mgr's ceph_module mocks to a dummy module
Basically, we add the initialization for the mgr's test suite to a
location that *all* mgr module test suites include in their PYTHONPATH

I'm not sure why this worked before, when it stopped exactly, and what
changed to make it stop. Yay python.

This shouldn't be copied into the CMAKE_INSTALL_DIR, but we should
double check the built packages to be 100% sure.