pkgbuild: add pkg.ceph-node-proxy
This is seemingly something to do with cephadm and (Dell?) hardware
metrics using the RedFishAPI... api. Apparently it was split out because
the upstream 'needed' it in v18 but the rest of the changes around it
from v19 weren't going to be backported.

Seems fairly worthless for ArchLinux but who can say.

References: https://github.com/ceph/ceph/commit/e1630fb8fcfa4460306b5e506c6a436f68c505a8