repo: add backport-ec-corruption-fix.patch
This is a fix for corruption of erasure coded pools which affects v18,
v19 and v20.

This has apparently been an issue from 19.2.0, but only if an EC pool
has allow_ec_overwrites enabled on it, which is not the default.

We pull the patch directly from the upstream backport PR.

References: https://github.com/ceph/ceph/pull/65963